Replaced two methods out of DB_FriendsList_Controller and moved them too DB_FriendsLists queryUserExists - returns 1 for true, 0 for false, -1 for error queryFriendExists - returns 1 for true, 0 for false, -1 for error
